powercms/requirements.txt
Django<4
mysqlclient==2.1.1
gunicorn==20.1.0

beautifulsoup4==4.12.0
cachetools==5.2.0
cchardet==2.1.7
certifi
chardet==5.0.0
django-admin-tools==0.9.2
django-ckeditor==5.9.0
django-js-asset==2.0.0
django-mptt-admin==2.3.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.17
django-smart-selects==1.6.0
django-storages==1.13.1
django-tabbed-admin==1.0.4
django-tagging==0.5.0
django-taggit==1.3.0
django-taggit-labels==0.7.0
easy-thumbnails==2.8.3
httplib2==0.20.4
idna==3.4

protobuf==4.21.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyPDF2==2.11.1
pytz==2020.1
reportlab==3.6.11
requests
rsa==4.9
six==1.16.0
sorl-thumbnail==12.9.0
soupsieve==2.3.2
sqlparse==0.4.3
swapper==1.3.0
uritemplate==4.1.1
urllib3
webencodings
django-recaptcha==3.0.0
django-compressor==4.3.1
django-crispy-forms==2.0
poweradmin @ git+https://bitbucket.org/josir/poweradmin.git@django32